requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
Django==3.1.1
django-cors-headers==3.5.0
django-safedelete==0.5.6
djangorestframework==3.11.1
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
